refactor: decouple interaction worker and modernize logger architecture
Worker changes:
- Split InteractionWorker into smaller, domain-specific files (worker, batch, redis, flush, health)
- Isolated event batching logic and Redis Stream operations
- Improved maintainability and testability of the worker package

Logger changes:
- Implemented Factory pattern for logger initialization
- Added dedicated Env type with support for Production, Development, and Test modes
- Disabled file logging in Test environment to keep CI/CD clean
- Separated interface implementation from configuration logic
