+++ internal/shared/logger/types.go
@@ -1,0 +1,9 @@
+package logger
+
+type Env string
+
+const (
+	EnvProduction  Env = "production"
+	EnvDevelopment Env = "development"
+	EnvTest        Env = "test"
+)
